Fix review notes.
